Added unit tests and fixed program bugs with these tests
